requirements.txt
python-telegram-bot <14
environs            <10
click               <9
ujson               <5
